+++ RockLib.Secrets.Aws/AwsSecretsConfigurationBuilderExtensions.cs
@@ -16,12 +16,11 @@ namespace RockLib.Secrets.Aws
         /// <param name="secretId">The Amazon Resource Name (ARN) or the friendly name of the secret.</param>
         /// <param name="secretKey">The key of the secret in AWS.</param>
         /// <param name="secretsManager">
-        /// The <see cref="IAmazonSecretsManager"/> client used for routing calls to AWS. If <see langword="null"/>,
-        /// then <see cref="AwsSecret.DefaultSecretsManager"/> is used instead.
+        /// The <see cref="IAmazonSecretsManager"/> client used for routing calls to AWS.
         /// </param>
         /// <returns>The same <see cref="ISecretsConfigurationBuilder"/>.</returns>
         public static ISecretsConfigurationBuilder AddAwsSecret(this ISecretsConfigurationBuilder builder,
-            string configurationKey, string secretId, string? secretKey = null, IAmazonSecretsManager? secretsManager = null)
+            string configurationKey, string secretId, string? secretKey, IAmazonSecretsManager secretsManager)
         {
             if (builder is null)
             {